# Optimize some loops with math/bits package