Benchmarks master:

```
go test -bench ValidMoves -benchmem
goos: windows
goarch: amd64
pkg: github.com/notnil/chess
BenchmarkValidMoves-12            200000              8303 ns/op            5304 B/op         69 allocs/op
BenchmarkValidMoves2-12           100000             18452 ns/op           12112 B/op        152 allocs/op
PASS
ok      github.com/notnil/chess 6.165s
```

This branch:

```
go test -bench ValidMoves -benchmem
goos: windows
goarch: amd64
pkg: github.com/notnil/chess
BenchmarkValidMoves-12            200000              7761 ns/op            5320 B/op         85 allocs/op
BenchmarkValidMoves2-12           100000             17485 ns/op           12160 B/op        170 allocs/op
PASS
ok      github.com/notnil/chess 6.073s
```

It isn't as dramatic as I had hoped, but eliminates looping over every square potentially multiple times per piece. Will continue benchmarking and profiling for further gains.

EDIT: With improved bitboard reverse:

```
go test -bench=ValidMoves -benchmem
goos: windows
goarch: amd64
pkg: github.com/notnil/chess
BenchmarkValidMoves-12            200000              6992 ns/op            5320 B/op         85 allocs/op
BenchmarkValidMoves2-12           100000             16368 ns/op           12160 B/op        170 allocs/op
PASS
ok      github.com/notnil/chess 5.382s
```