@captncraig 2018-12-28
![profile001](https://user-images.githubusercontent.com/192540/50499023-e2f6eb00-0a13-11e9-9787-155a2712cbcc.png)

Profile for future reference. Looks like I'll dissect addTags next.