@notnil 2020-02-23
I eventually added math/bits for newer go versions.  Closing.